pyproject.toml
[tool.poetry]
name = "gen-ai-intensive"
version = "0.1.0"
description = "Minimal setup for Google's 5-Day Gen AI Intensive course"
authors = ["Jason Walsh <j@wal.sh>"]
readme = "README.org"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
google-genai = "1.7.0"
python-dotenv = "^1.0.1"
hy = "^1.0"
tqdm = "^4.66.2"
langgraph = "0.3.21"
langchain-google-genai = "2.1.2"
langgraph-prebuilt = "0.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"